Add gp2040-CE gpio config files
This sets the correct pins for the front panel buttons since I used the SHARE one as a i2c pin for the oled. layout ends as START/SELECT/HOME/L3/R3.
Also sets the OLED's SDA/SCL pins and enables the function on the screen display tab of the webconfig.